+++ sphinx/source/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 9.2b1
 -----
 * Added support to :ref:`tap2sna.py` for TZX block type 0x15 (direct recording)
+* :ref:`tapinfo.py` now shows info for TZX block type 0x15 (direct recording)
 
 9.1 (2024-02-03)
 ----------------